requirements.txt
Flask==1.1.2
line-bot-sdk==1.16.0
gunicorn
apscheduler==3.6.3
googletrans==3.1.0a0
mplfinance
pandas_datareader== 0.10.0
pyimgur
twstock
lxml